refactor to shared, rename lib to src
