+++ .npmignore
@@ -1,4 +1,4 @@
-lib/
+src/
 tests/
 tsconfig.json
 jasmine.json
