+++ src/collectionFactory.ts
