+++ src/documentException.ts
