+++ tests/documentSpec.ts
@@ -1,7 +1,7 @@
 import { ObjectID } from 'mongodb'
 import { MongoMemoryServer } from 'mongodb-memory-server'
 
-import { close, connect } from '../lib/index'
+import { close, connect } from '../src/index'
 import { IUser, User, UserCollection } from './user'
 
 let mongoServerInstance: MongoMemoryServer
